requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
wandb>=0.15.0
tiktoken>=0.5.0
tqdm>=4.65.0
numpy>=1.24.0
pyyaml>=6.0.0
